third_party/blink/web_tests/external/wpt/svg/idlharness.window-expected.txt

This is a testharness.js-based test.
Found 182 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] SVGElement interface: attribute correspondingElement
  assert_true: The prototype object must have a property "correspondingElement" expected true got false
[FAIL] SVGElement interface: attribute correspondingUseElement
  assert_true: The prototype object must have a property "correspondingUseElement" expected true got false
[FAIL] SVGGeometryElement interface: operation isPointInFill(optional DOMPointInit)
  assert_equals: property has wrong .length expected 0 but got 1
[FAIL] SVGGeometryElement interface: operation isPointInStroke(optional DOMPointInit)
  assert_equals: property has wrong .length expected 0 but got 1
[FAIL] SVGAnimatedRect interface: objects.svg.viewBox must inherit property "baseVal" with the proper type
  assert_true: wrong type: not a DOMRect expected true got false
[FAIL] SVGAnimatedRect interface: objects.svg.viewBox must inherit property "animVal" with the proper type
  assert_true: wrong type: not a DOMRectReadOnly expected true got false
[FAIL] SVGSVGElement interface: operation createSVGTransformFromMatrix(optional DOMMatrix2DInit)
  assert_equals: property has wrong .length expected 0 but got 1
[FAIL] SVGSVGElement interface: objects.svg must inherit property "currentTranslate" with the proper type
  assert_true: wrong type: not a DOMPointReadOnly expected true got false
[FAIL] SVGElement interface: objects.svg must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.svg must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.g must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.g must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.defs must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.defs must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.desc must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.desc must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.metadata must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.metadata must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.title must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.title must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGSymbolElement interface: existence and properties of interface object
  assert_equals: prototype of SVGSymbolElement is not SVGGraphicsElement expected function "function SVGGraphicsElement() { [native code] }" but got function "function SVGElement() { [native code] }"
[FAIL] SVGSymbolElement interface: existence and properties of interface prototype object
  assert_equals: prototype of SVGSymbolElement.prototype is not SVGGraphicsElement.prototype expected object "[object SVGGraphicsElement]" but got object "[object SVGElement]"
[FAIL] SVGGraphicsElement interface: objects.symbol must inherit property "transform" with the proper type
  assert_inherits: property "transform" not found in prototype chain
[FAIL] SVGGraphicsElement interface: objects.symbol must inherit property "getBBox(optional SVGBoundingBoxOptions)" with the proper type
  assert_inherits: property "getBBox" not found in prototype chain
[FAIL] SVGGraphicsElement interface: calling getBBox(optional SVGBoundingBoxOptions) on objects.symbol with too few arguments must throw TypeError
  assert_inherits: property "getBBox" not found in prototype chain
[FAIL] SVGGraphicsElement interface: objects.symbol must inherit property "getCTM()" with the proper type
  assert_inherits: property "getCTM" not found in prototype chain
[FAIL] SVGGraphicsElement interface: objects.symbol must inherit property "getScreenCTM()" with the proper type
  assert_inherits: property "getScreenCTM" not found in prototype chain
[FAIL] SVGGraphicsElement interface: objects.symbol must inherit property "requiredExtensions" with the proper type
  assert_inherits: property "requiredExtensions" not found in prototype chain
[FAIL] SVGGraphicsElement interface: objects.symbol must inherit property "systemLanguage" with the proper type
  assert_inherits: property "systemLanguage" not found in prototype chain
[FAIL] SVGElement interface: objects.symbol must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.symbol must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGUseElement interface: attribute instanceRoot
  assert_true: The prototype object must have a property "instanceRoot" expected true got false
[FAIL] SVGUseElement interface: attribute animatedInstanceRoot
  assert_true: The prototype object must have a property "animatedInstanceRoot" expected true got false
[FAIL] SVGUseElement interface: objects.use must inherit property "instanceRoot" with the proper type
  assert_inherits: property "instanceRoot" not found in prototype chain
[FAIL] SVGUseElement interface: objects.use must inherit property "animatedInstanceRoot" with the proper type
  assert_inherits: property "animatedInstanceRoot" not found in prototype chain
[FAIL] SVGElement interface: objects.use must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.use must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGUseElementShadowRoot interface: existence and properties of interface object
  assert_own_property: self does not have own property "SVGUseElementShadowRoot" expected property "SVGUseElementShadowRoot" missing
[FAIL] SVGUseElementShadowRoot interface object length
  assert_own_property: self does not have own property "SVGUseElementShadowRoot" expected property "SVGUseElementShadowRoot" missing
[FAIL] SVGUseElementShadowRoot interface object name
  assert_own_property: self does not have own property "SVGUseElementShadowRoot" expected property "SVGUseElementShadowRoot" missing
[FAIL] SVGUseElementShadowRoot interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "SVGUseElementShadowRoot" expected property "SVGUseElementShadowRoot" missing
[FAIL] SVGUseElementShadowRoot interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "SVGUseElementShadowRoot" expected property "SVGUseElementShadowRoot" missing
[FAIL] SVGUseElementShadowRoot interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "SVGUseElementShadowRoot" expected property "SVGUseElementShadowRoot" missing
[FAIL] ShadowAnimation interface: existence and properties of interface object
  assert_own_property: self does not have own property "ShadowAnimation" expected property "ShadowAnimation" missing
[FAIL] ShadowAnimation interface object length
  assert_own_property: self does not have own property "ShadowAnimation" expected property "ShadowAnimation" missing
[FAIL] ShadowAnimation interface object name
  assert_own_property: self does not have own property "ShadowAnimation" expected property "ShadowAnimation" missing
[FAIL] ShadowAnimation interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "ShadowAnimation" expected property "ShadowAnimation" missing
[FAIL] ShadowAnimation interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "ShadowAnimation" expected property "ShadowAnimation" missing
[FAIL] ShadowAnimation interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "ShadowAnimation" expected property "ShadowAnimation" missing
[FAIL] ShadowAnimation interface: attribute sourceAnimation
  assert_own_property: self does not have own property "ShadowAnimation" expected property "ShadowAnimation" missing
[FAIL] SVGElement interface: objects.switch must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.switch must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.style must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.style must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGTransform interface: operation setMatrix(optional DOMMatrix2DInit)
  assert_equals: property has wrong .length expected 0 but got 1
[FAIL] SVGTransform interface: objects.svg.createSVGTransform() must inherit property "matrix" with the proper type
  assert_true: wrong type: not a DOMMatrix expected true got false
[FAIL] SVGTransformList interface: operation createSVGTransformFromMatrix(optional DOMMatrix2DInit)
  assert_equals: property has wrong .length expected 0 but got 1
[FAIL] SVGElement interface: objects.path must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.path must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.rect must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.rect must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.circle must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.circle must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.ellipse must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.ellipse must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.line must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.line must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.polyline must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.polyline must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.polygon must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.polygon must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGTextContentElement interface: operation getCharNumAtPosition(optional DOMPointInit)
  assert_equals: property has wrong .length expected 0 but got 1
[FAIL] SVGElement interface: objects.text must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.text must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.tspan must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.tspan must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.textPath must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.textPath must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.image must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.image must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.foreignObject must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.foreignObject must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGMarkerElement interface: attribute orient
  assert_true: The prototype object must have a property "orient" expected true got false
[FAIL] SVGMarkerElement interface: objects.marker must inherit property "orient" with the proper type
  assert_inherits: property "orient" not found in prototype chain
[FAIL] SVGElement interface: objects.marker must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.marker must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.linearGradient must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.linearGradient must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.radialGradient must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.radialGradient must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.stop must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.stop must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.pattern must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.pattern must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGScriptElement interface: attribute crossOrigin
  assert_true: The prototype object must have a property "crossOrigin" expected true got false
[FAIL] SVGScriptElement interface: objects.script must inherit property "crossOrigin" with the proper type
  assert_inherits: property "crossOrigin" not found in prototype chain
[FAIL] SVGElement interface: objects.script must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.script must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGAElement interface: attribute download
  assert_true: The prototype object must have a property "download" expected true got false
[FAIL] SVGAElement interface: attribute ping
  assert_true: The prototype object must have a property "ping" expected true got false
[FAIL] SVGAElement interface: attribute rel
  assert_true: The prototype object must have a property "rel" expected true got false
[FAIL] SVGAElement interface: attribute relList
  assert_true: The prototype object must have a property "relList" expected true got false
[FAIL] SVGAElement interface: attribute hreflang
  assert_true: The prototype object must have a property "hreflang" expected true got false
[FAIL] SVGAElement interface: attribute type
  assert_true: The prototype object must have a property "type" expected true got false
[FAIL] SVGAElement interface: attribute text
  assert_true: The prototype object must have a property "text" expected true got false
[FAIL] SVGAElement interface: attribute referrerPolicy
  assert_true: The prototype object must have a property "referrerPolicy" expected true got false
[FAIL] SVGAElement interface: attribute origin
  assert_true: The prototype object must have a property "origin" expected true got false
[FAIL] SVGAElement interface: attribute protocol
  assert_true: The prototype object must have a property "protocol" expected true got false
[FAIL] SVGAElement interface: attribute username
  assert_true: The prototype object must have a property "username" expected true got false
[FAIL] SVGAElement interface: attribute password
  assert_true: The prototype object must have a property "password" expected true got false
[FAIL] SVGAElement interface: attribute host
  assert_true: The prototype object must have a property "host" expected true got false
[FAIL] SVGAElement interface: attribute hostname
  assert_true: The prototype object must have a property "hostname" expected true got false
[FAIL] SVGAElement interface: attribute port
  assert_true: The prototype object must have a property "port" expected true got false
[FAIL] SVGAElement interface: attribute pathname
  assert_true: The prototype object must have a property "pathname" expected true got false
[FAIL] SVGAElement interface: attribute search
  assert_true: The prototype object must have a property "search" expected true got false
[FAIL] SVGAElement interface: attribute hash
  assert_true: The prototype object must have a property "hash" expected true got false
[FAIL] SVGAElement interface: objects.a must inherit property "download" with the proper type
  assert_inherits: property "download" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "ping" with the proper type
  assert_inherits: property "ping" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "rel" with the proper type
  assert_inherits: property "rel" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "relList" with the proper type
  assert_inherits: property "relList" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "hreflang" with the proper type
  assert_inherits: property "hreflang" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "type" with the proper type
  assert_inherits: property "type" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "text" with the proper type
  assert_inherits: property "text" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "referrerPolicy" with the proper type
  assert_inherits: property "referrerPolicy" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "origin" with the proper type
  assert_inherits: property "origin" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "protocol" with the proper type
  assert_inherits: property "protocol" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "username" with the proper type
  assert_inherits: property "username" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "password" with the proper type
  assert_inherits: property "password" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "host" with the proper type
  assert_inherits: property "host" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "hostname" with the proper type
  assert_inherits: property "hostname" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "port" with the proper type
  assert_inherits: property "port" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "pathname" with the proper type
  assert_inherits: property "pathname" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "search" with the proper type
  assert_inherits: property "search" not found in prototype chain
[FAIL] SVGAElement interface: objects.a must inherit property "hash" with the proper type
  assert_inherits: property "hash" not found in prototype chain
[FAIL] SVGElement interface: objects.a must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.a must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.view must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.view must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] TimeEvent interface: existence and properties of interface object
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] TimeEvent interface object length
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] TimeEvent interface object name
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] TimeEvent interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] TimeEvent interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] TimeEvent interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] TimeEvent interface: attribute view
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] TimeEvent interface: attribute detail
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] TimeEvent interface: operation initTimeEvent(DOMString, Window?, long)
  assert_own_property: self does not have own property "TimeEvent" expected property "TimeEvent" missing
[FAIL] SVGElement interface: objects.animate must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.animate must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.set must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.set must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.animateMotion must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.animateMotion must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.mpath must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.mpath must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGElement interface: objects.animateTransform must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.animateTransform must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
[FAIL] SVGDiscardElement interface: existence and properties of interface object
  assert_own_property: self does not have own property "SVGDiscardElement" expected property "SVGDiscardElement" missing
[FAIL] SVGDiscardElement interface object length
  assert_own_property: self does not have own property "SVGDiscardElement" expected property "SVGDiscardElement" missing
[FAIL] SVGDiscardElement interface object name
  assert_own_property: self does not have own property "SVGDiscardElement" expected property "SVGDiscardElement" missing
[FAIL] SVGDiscardElement interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "SVGDiscardElement" expected property "SVGDiscardElement" missing
[FAIL] SVGDiscardElement interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "SVGDiscardElement" expected property "SVGDiscardElement" missing
[FAIL] SVGDiscardElement interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "SVGDiscardElement" expected property "SVGDiscardElement" missing
[FAIL] SVGDiscardElement must be primary interface of objects.discard
  assert_own_property: self does not have own property "SVGDiscardElement" expected property "SVGDiscardElement" missing
[FAIL] Stringification of objects.discard
  assert_class_string: class string of objects.discard expected "[object SVGDiscardElement]" but got "[object SVGElement]"
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "targetElement" with the proper type
  assert_inherits: property "targetElement" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "onbegin" with the proper type
  assert_inherits: property "onbegin" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "onend" with the proper type
  assert_inherits: property "onend" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "onrepeat" with the proper type
  assert_inherits: property "onrepeat" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "getStartTime()" with the proper type
  assert_inherits: property "getStartTime" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "getCurrentTime()" with the proper type
  assert_inherits: property "getCurrentTime" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "getSimpleDuration()" with the proper type
  assert_inherits: property "getSimpleDuration" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "beginElement()" with the proper type
  assert_inherits: property "beginElement" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "beginElementAt(float)" with the proper type
  assert_inherits: property "beginElementAt" not found in prototype chain
[FAIL] SVGAnimationElement interface: calling beginElementAt(float) on objects.discard with too few arguments must throw TypeError
  assert_inherits: property "beginElementAt" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "endElement()" with the proper type
  assert_inherits: property "endElement" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "endElementAt(float)" with the proper type
  assert_inherits: property "endElementAt" not found in prototype chain
[FAIL] SVGAnimationElement interface: calling endElementAt(float) on objects.discard with too few arguments must throw TypeError
  assert_inherits: property "endElementAt" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "requiredExtensions" with the proper type
  assert_inherits: property "requiredExtensions" not found in prototype chain
[FAIL] SVGAnimationElement interface: objects.discard must inherit property "systemLanguage" with the proper type
  assert_inherits: property "systemLanguage" not found in prototype chain
[FAIL] SVGElement interface: objects.discard must inherit property "correspondingElement" with the proper type
  assert_inherits: property "correspondingElement" not found in prototype chain
[FAIL] SVGElement interface: objects.discard must inherit property "correspondingUseElement" with the proper type
  assert_inherits: property "correspondingUseElement" not found in prototype chain
Harness: the test ran to completion.
